+++ resources/views/tasks/show.blade.php
@@ -10,19 +10,21 @@
         </tr>
         <tr>
             <th>タスク内容</th>
-            <td>{{$task->content}}</td>
+            <td>{!! nl2br(e($task->content)) !!}</td>
         </tr>
         <tr>
             <th>進捗状況</th>
-            <td>{{$task->status}}</td>
+            <td>{!! nl2br(e($task->status)) !!}</td>
         </tr>
     </table>
     
     {{--タスク編集ページへのリンク--}}
     {!! link_to_route("tasks.edit", "このタスクを編集", ["task"=>$task->id], ["class"=>"btn btn-light"]) !!}
     
-    {{--タスクの削除--}}
-    {!! Form::model($task, ["route"=>["tasks.destroy", $task->id], "method"=>"delete"]) !!}
-        {!! Form::submit("削除", ["class"=>"btn btn-danger"]) !!}
-    {!! Form::close() !!}
+    @if(Auth::id()==$task->user_id)
+        {{--タスクの削除--}}
+        {!! Form::model($task, ["route"=>["tasks.destroy", $task->id], "method"=>"delete"]) !!}
+            {!! Form::submit("削除", ["class"=>"btn btn-danger"]) !!}
+        {!! Form::close() !!}
+    @endif
 @endsection
